project/requirements-final.txt
streamlit==1.24.0
google-generativeai==0.3.1
elevenlabs==0.2.24
python-dotenv==1.0.0
Pillow==9.5.0
requests==2.31.0
python-dateutil==2.8.2